Remove RPC connection for channel ApplyMove.
This removes the RPC connection argument that is passed to the game
channel ApplyMove method.  The reason for passing it was so it can be
used to verify signatures; but this is a bad reason to code something
as complex as a JSON-RPC client into the very simple and pure BoardRules
interface.

Actually no code in production is using it at the moment, so removing
the argument is easy and simplifies some things (e.g. in tests).  We are
going to replace the RPC for signature verification with a generic and
simpler "signer/verifier" interface as well, so this will be the proper
interface for rules that need signatures.